ext: lib: mbedtls: Fix MBEDTLS_PEM_PARSE_C definition in generic config
During recent refactoring of mbedTLS generic config file, a regression
slipped in that prevented MBEDTLS_PEM_PARSE_C from being set, even
though the option was selected in Kconfig. The reason for this is the
fact that this config has dependency to MBEDTLS_X509_CRT_PARSE_C but it
was moved above the line where MBEDTLS_X509_CRT_PARSE_C was actually
set. Therefore, this dependency was never satisfied.

Signed-off-by: Robert Lubos <robert.lubos@nordicsemi.no>