cmake: Misc. cleanup of build scripts
Misc. cleanup of build scripts. Most importantly, mbedTLS is no longer
linked with zephyr_interface. It should not be as there is no reason
to do so. This aligns with
https://github.com/zephyrproject-rtos/zephyr/pull/19346

Also, glob the source files in 'library' instead of listing them to
make it clear that all files are used.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>