Revert "Require framework directory to exist when building"
This reverts commit 469f7811fac98617e544aac2b1b2db56040c95db.

Remove the dependency on the framework directory.

It was introduced in MbedTLS 3.6.0 and for now
doesn't do much more than make CMake fail if
`git submodule update --init` is not run.

This needs to be revisited once bigger changes
are made (with Mbed TLS 4.0) and the whole
repository structure changes.

Signed-off-by: Tomi Fontanilles <tomi.fontanilles@nordicsemi.no>